refactor(web): login features->entities 폴더 이전
